Local mode (#84)
* Test inter-frame communication to Web service

There is code in the extensions-hello-world repository to test this.

* Move the test endpoint handler into start.js

This puts the code that will respond to event publication requests with
the code that will send those events to subscribers.

* Move application configuration into start.js

* Remove coordinator from the top-level HTML

Replace it with an ad-hoc object that has all necessary fields.

* Locally source coordinator, twitch-ext, supervisor

* Replace pubsub references with localhost

* Implement local onAuthorized

* Add stubs for other functions except PubSub

Several of these are null and will fail if invoked.

* Implement a local pubsub

* Create endpoints for API in local mode

* Send extension messages to all clients

* Implement other non-PubSub local functions

* Implement PubSub local functions

* Remove authentication endpoint

* Restore coordinator in the top-level HTML

* Start local mode if requested on command line

* Enable local hosting of front-end

* Support local-online switching

* Create a script to create an extension manifest

* Remove magic numbers and strings

* Deobfuscate supervisor.js

* Complete stub implementation of twitch-ext.js

* Move local mode code into a separate file

* Address initial comments

* First Pass of Documentation for Local Mode

* Responded to David's Feedback

Grammar fixes, added some updates around pathing of filenames.

* Update the read-me file

* Improve manifest generation

* Additional Documentation Edits

Adding clarity to getting started guide and starting on more detail for local manifest and editing the run list.

* Added description on editing manifest.json file

Also did some cleanup on the first set up steps.

* Generated readme.html

Made some readme changes to better convert to HTML in a more readable way.

* Remove references to run list in documentation

Pulled out the bits about the runlist  - will bring back in next week when it's introduced.

* Adding Blurb About Local Mode Secret

Adding how to use the local mode secret in documentation.   Also updated the README.html for this information as well.