Realize the concept of a Project (#180)
* Implement projects

* Copy SSL certificates

* Perform back-end command variable substitutions

* Update projects for local mode

* Conditionally allow pre-existing project folder

* Conditionally allow empty project folders

* Use BEM to style the project view

* Use BEM to style the "create project" dialog

* Create the project drop-down menu

* Move extension views into the project

* Shut down hosting, if necessary

* Create a property for the current project index

* Improve front-end hosting

It now creates a cross-platform implementation that doesn't rely on a
shell to execute the command.  As a result, the child process can be
killed instead of only the spawning shell.

* Improve back-end hosting

With the assistance of a newly-required library, this implementation
mimics that of the front-end hosting.

* Stop hosting when creating a project

* Implement the mock trigger bar

* Implement mock trigger messaging

* Move mock trigger bar to extension view container

* Open tabs for documentation and tutorial

* Provide error feedback in "create project" dialog

* Implement explicitly stopping hosting

* Disable project name textbox for online extensions

For online extensions, the name is taken from the extension.  There is a
tooltip to describe the situation.

* Remove installation and configuration of Git

* Simplify the "run" script

* Add iconUrls field to ExtensionObject class

* Remove the extension manifest from local storage

This will more easily allow updating the manifest.

* Refresh online manifest if requested

* Remove "Generate Scaffolding" option

No scaffolding projects are yet available.

* Cut Down the Read Me to Just Quick Start and Trouble Shooting

Since we are moving our docs onto the dev site, the readme has been significantly simplified.

* Replace HTML-based README with PDF

Updating the documentation reflect the new state of the world

* Add Windows 10 troubleshooting

Line discusses needing to have your security settings to Developer Mode

* Automatically scroll the console as necessary

* Replace trigger switch HTML with SVG

* Create platform-agnostic back-end command

* Implement custom front-end hosting

* Remove install and configure from run

The run scripts will issue a warning and fail to run if the
configuration is not valid.

* Create a single certificate for both DNS entries

Use both localhost and localhost.rig.twitch.tv as DNS entries in the
generated certificate.  This certificate is good for hosting the front-
and back-end of an extension in either domain.

* Default extension visibility to true

* Create project folder only, if requested

* Use React ref in the console

* Protect against missing client ID

* Use spawn instead of spawnSync

Using spawnSync causes the Node event loop to hang.  This disrupts the
networking sufficiently to cause random errors (e.g. client-side network
error and mysterious restarting of a POST request).

* Remove unnecessary import directives