Address issues with Configuration Service changes (#189)
* Include Developer Rig request header

* Create a separate payload type for each token type

* Set iframe onload instead of using keys

The purpose of the keys was to ensure the enclosed iframes were
recreated so they had their onload event set.  Setting iframe onload
upon binding instead of component mounting ensures the desired behavior,
obviating the need for keys.

* Add API tests for Configuration Service

* Improve channel wording

Either a channel ID or a channel name is acceptable.  In case of the
latter, the Developer Rig will fetch the former.

* Remove extraneous API error messages

The API-provided error messages are sufficient.